# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShiftMind API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approvals": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Create Approval",
        "operationId": "create_approval_api_v1_approvals_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "List Approvals",
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "name": "schedule_run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalListOut"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/decision": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Decide Approval Route",
        "operationId": "decide_approval_route_api_v1_approvals__approval_id__decision_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "Get Approval",
        "operationId": "get_approval_api_v1_approvals__approval_id__get",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsV1"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApprovalDecisionIn": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision"
          },
          "expected_resource_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Resource Version"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "expected_resource_version"
        ],
        "title": "ApprovalDecisionIn"
      },
      "ApprovalListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ApprovalListOut"
      },
      "ApprovalOut": {
        "properties": {
          "approval_id": {
            "type": "string",
            "format": "uuid",
            "title": "Approval Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "consumed",
              "rejected",
              "expired",
              "stale"
            ],
            "title": "State"
          },
          "schedule_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Schedule Run Id"
          },
          "candidate_schedule_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Schedule Version Id"
          },
          "baseline_schedule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Schedule Version"
          },
          "scenario_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Version Id"
          },
          "consequence_summary": {
            "type": "string",
            "title": "Consequence Summary"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "agent_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "resource_version": {
            "type": "integer",
            "title": "Resource Version"
          }
        },
        "type": "object",
        "required": [
          "approval_id",
          "state",
          "schedule_run_id",
          "candidate_schedule_version_id",
          "baseline_schedule_version",
          "scenario_version_id",
          "consequence_summary",
          "policy_version",
          "agent_run_id",
          "created_at",
          "expires_at",
          "resource_version"
        ],
        "title": "ApprovalOut"
      },
...
      "ApprovalRequestIn": {
        "properties": {
          "schedule_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Schedule Run Id"
          },
          "expected_resource_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Resource Version"
          },
          "expected_baseline_schedule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Baseline Schedule Version"
          }
        },
        "type": "object",
        "required": [
          "schedule_run_id",
          "expected_resource_version",
          "expected_baseline_schedule_version"
        ],
        "title": "ApprovalRequestIn"
      },
...
      "ProblemDetailsV1": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "expected": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected"
          },
          "current": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetailsV1"
      },
...
    }
  }
}